Removed equal overrides in POCO contract classes
The POCO contract classes previously contained both Equals and GetHashCode
method overrides to allow test code to call Assert.AreEqual(). The problem
with this approach is that the contracts may not want to perform as
exhaustive of equality comparison as test code.

To fix this, the overrides have been removed and the CompareNetObjects
project has been added. This project allows deep comparison of objects.
Now the tests can perform deep equality checks while keeping the contracts
prestine with their own appropriate equality logic.